Changed navbar so that we can now scroll on page while it stays in place. Also did some formatting on the homepage

+++ Client-Ui/src/Components/CryptoCard.js
@@ -47,7 +47,7 @@ export default function CryptoCard(props) {
   };
 
   const getChangeColor = (changeValue) => {
-    if (parseInt(changeValue) < 0) {
+    if (parseFloat(changeValue) < 0) {
       return "red";
     } else {
       return "green";
